backend/web/requirements.txt
Flask
sqlalchemy
psycopg2-binary
alembic>=1.3
apscheduler
flask_caching
requests
PyYaml
